requirements.txt
# Core GUI
PyQt5==5.15.10
PyQt5-sip==12.13.0

# Audio Processing
sounddevice==0.5.0
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
webrtcvad==2.0.10

# Speech Recognition
openai-whisper==20231117

# Translation
googletrans==4.0.0rc1

# Text-to-Speech (Windows compatible)
pyttsx3==2.90
pywin32==306
gTTS==2.5.0
pygame==2.5.2
edge-tts==6.1.9

# Machine Learning
scikit-learn==1.3.2

# NLP
thefuzz==0.22.1
python-Levenshtein==0.25.0
nltk==3.8.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2

# System Utilities
psutil==6.0.0
pyperclip==1.9.0

# Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9

# Automation
selenium==4.15.2
webdriver-manager==4.0.1
pyautogui==0.9.54
pillow==10.2.0

# Windows Notifications
win10toast==0.9

# HTTP Requests
requests==2.31.0
